backend/requirements.txt
Flask>=2.0.0
python-dotenv>=0.19.0
openai>=1.0.0
google-generativeai>=0.8.0
anthropic>=0.25.0
requests>=2.28.0
flask_cors>=4.0.0
